+++ src/commands/serve.rs
@@ -1,3 +1,5 @@
+use crate::commands::bundle::bundle;
+
 use axum::{
     extract::Query,
     response::{IntoResponse, Response},
@@ -11,6 +13,15 @@ use tower::ServiceBuilder;
 use tower_http::trace::{self, TraceLayer};
 use tracing::{info, Level};
 use uuid::Uuid;
+
+use anyhow::Result;
+use std::path::Path;
+
+use notify::event::{DataChange, EventKind, ModifyKind};
+use notify::{Event, RecommendedWatcher, RecursiveMode, Watcher};
+use tokio::sync::mpsc;
+
+
 
 #[derive(Deserialize)]
 struct FlowQuery {
@@ -95,7 +106,31 @@ async fn sessions() -> Json<SessionResponse> {
     })
 }
 
-pub async fn serve() -> Result<(), Box<dyn std::error::Error>> {
+async fn watch(config_path: &str) -> notify::Result<()> {
+    let (tx, mut rx) = mpsc::channel::<Event>(100);
+
+    let mut watcher: RecommendedWatcher = notify::recommended_watcher(move |res| {
+        if let Ok(event) = res {
+            let _ = tx.try_send(event);
+        } else if let Err(e) = res {
+            eprintln!("Watch error: {:?}", e);
+        }
+    })?;
+
+    watcher.watch(Path::new("src"), RecursiveMode::Recursive)?;
+    watcher.watch(Path::new(config_path), RecursiveMode::NonRecursive)?;
+    println!("Watching all files in 'src' and '{}'", config_path);
+
+    while let Some(_event) = rx.recv().await {
+        if _event.kind == EventKind::Modify(ModifyKind::Data(DataChange::Content)) {
+            let _ = bundle(config_path);
+        }
+    }
+
+    Ok(())
+}
+
+pub async fn serve(config_path: &str, should_watch: &bool) -> Result<(), Box<dyn std::error::Error>> {
     let port = 8080;
     let addr = SocketAddr::from(([0, 0, 0, 0], port));
 
@@ -121,8 +156,21 @@ pub async fn serve() -> Result<(), Box<dyn std::error::Error>> {
 
     info!("Listening on port {}...", port);
 
-    let listener = tokio::net::TcpListener::bind(addr).await?;
-    axum::serve(listener, app.into_make_service()).await?;
+    if *should_watch {
+        tokio::try_join!(
+            async {
+                let listener = tokio::net::TcpListener::bind(addr).await?;
+                axum::serve(listener, app.into_make_service()).await?;
+                Ok::<_, Box<dyn std::error::Error>>(())
+            },
+            async {
+                watch(config_path).await.map_err(|e| Box::new(e) as Box<dyn std::error::Error>)
+            }
+        )?;
+    } else {
+        let listener = tokio::net::TcpListener::bind(addr).await?;
+        axum::serve(listener, app.into_make_service()).await?;
+    }
 
     Ok(())
 }
